reformat the files hierarchy, change click location in Nearby test
