+++ app/src/main/java/ch/epfl/favo/auth/SignInActivity.java
@@ -22,8 +22,8 @@ import java.util.concurrent.CompletableFuture;
 import ch.epfl.favo.BuildConfig;
 import ch.epfl.favo.MainActivity;
 import ch.epfl.favo.R;
-import ch.epfl.favo.common.FavoLocation;
-import ch.epfl.favo.map.Locator;
+import ch.epfl.favo.gps.FavoLocation;
+import ch.epfl.favo.gps.Locator;
 import ch.epfl.favo.user.User;
 import ch.epfl.favo.user.UserUtil;
 import ch.epfl.favo.util.DependencyFactory;
