--- app/src/main/java/ch/epfl/favo/common/FavorNoLongerAvailableException.java
@@ -1,7 +1,0 @@
-package ch.epfl.favo.common;
-
-public class FavorNoLongerAvailableException extends RuntimeException {
-  public FavorNoLongerAvailableException(String s) {
-    super(s);
-  }
-}
