--- app/src/main/java/ch/epfl/favo/common/NotImplementedException.java
@@ -1,8 +1,0 @@
-package ch.epfl.favo.common;
-
-public class NotImplementedException extends RuntimeException {
-
-  public NotImplementedException() {
-    super();
-  }
-}
