+++ app/src/main/java/ch/epfl/favo/database/CollectionWrapper.java
@@ -1,4 +1,4 @@
-package ch.epfl.favo.common;
+package ch.epfl.favo.database;
 
 import android.location.Location;
