+++ app/src/main/java/ch/epfl/favo/exception/FavorNoLongerAvailableException.java
@@ -1,0 +1,7 @@
+package ch.epfl.favo.exception;
+
+public class FavorNoLongerAvailableException extends RuntimeException {
+  public FavorNoLongerAvailableException(String s) {
+    super(s);
+  }
+}
