+++ app/src/main/java/ch/epfl/favo/exception/NoPermissionGrantedException.java
@@ -1,4 +1,4 @@
-package ch.epfl.favo.common;
+package ch.epfl.favo.exception;
 
 public class NoPermissionGrantedException extends RuntimeException {
   public NoPermissionGrantedException(String s) {
