+++ app/src/main/java/ch/epfl/favo/exception/NotImplementedException.java
@@ -1,0 +1,8 @@
+package ch.epfl.favo.exception;
+
+public class NotImplementedException extends RuntimeException {
+
+  public NotImplementedException() {
+    super();
+  }
+}
