+++ app/src/main/java/ch/epfl/favo/favor/FavorUtil.java
@@ -12,8 +12,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
-import ch.epfl.favo.common.DatabaseUpdater;
-import ch.epfl.favo.common.NotImplementedException;
+import ch.epfl.favo.database.DatabaseUpdater;
+import ch.epfl.favo.exception.NotImplementedException;
 import ch.epfl.favo.util.DependencyFactory;
 
 /*
