+++ app/src/main/java/ch/epfl/favo/user/User.java
@@ -5,8 +5,8 @@ import android.location.Location;
 import java.util.Date;
 import java.util.Map;
 
-import ch.epfl.favo.common.Document;
-import ch.epfl.favo.common.FavoLocation;
+import ch.epfl.favo.database.Document;
+import ch.epfl.favo.gps.FavoLocation;
 
 /**
  * This class contains all the relevant information about users TODO: It should implement parcelable
