+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -27,7 +27,7 @@
         android:label="Chat" />
     <fragment
         android:id="@+id/nav_map"
-        android:name="ch.epfl.favo.view.tabs.MapsPage"
+        android:name="ch.epfl.favo.view.tabs.MapPage"
         android:label="Map"
         tools:layout="@layout/fragment_map">
         <action
