+++ app/src/sharedTestUtils/java/ch/epfl/favo/view/MockDatabaseWrapper.java
@@ -10,8 +10,8 @@ import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
 import ch.epfl.favo.FakeItemFactory;
-import ch.epfl.favo.common.DatabaseUpdater;
-import ch.epfl.favo.common.Document;
+import ch.epfl.favo.database.DatabaseUpdater;
+import ch.epfl.favo.database.Document;
 
 public class MockDatabaseWrapper<T extends Document> implements DatabaseUpdater<T> {
