+++ app/src/test/java/ch/epfl/favo/gps/FavoLocationTest.java
@@ -1,8 +1,10 @@
-package ch.epfl.favo.common;
+package ch.epfl.favo.gps;
 
 import android.location.Location;
 
 import org.junit.Test;
+
+import ch.epfl.favo.gps.FavoLocation;
 
 import static org.junit.Assert.assertEquals;
